# server/apiSpec.json
{
  "openapi": "3.0.0",
  "tags": [
    { "name": "Mutational Signature" },
    { "name": "Mutational Spectrum" },
    { "name": "Mutational Activity" },
    { "name": "Mutational Association" },
    { "name": "Etiology" }
  ],
  "paths": {
    "/api/mutational_spectrum_options": {
      "get": {
        "tags": ["Mutational Spectrum"],
        "summary": "Retrieves seqmatrix options",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select which study to query data from.",
            "required": false,
            "schema": { "type": "string" },
            "value": "PCAWG"
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Select cancer belonging to the study",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "sample",
            "in": "query",
            "description": "Enter sample.",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "profile",
            "in": "query",
            "description": "Select a profile type to filter data by.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["", "SBS", "DBS", "ID"]
            },
            "value": ""
          },
          {
            "name": "matrix",
            "in": "query",
            "description": "Select a matrix size that coresponds to the profile.",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_spectrum": {
      "get": {
        "tags": ["Mutational Spectrum"],
        "summary": "Retrieves seqmatrix data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select which study to query data from.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "value": "PCAWG"
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Select cancer belonging to the study",
            "required": true,
            "schema": {},
            "value": "Lung-AdenoCA"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "value": "WGS"
          },
          {
            "name": "sample",
            "in": "query",
            "description": "Enter sample.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "SP50263"
          },
          {
            "name": "profile",
            "in": "query",
            "description": "Select a profile type to filter data by.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["SBS", "DBS", "ID"]
            },
            "value": "SBS"
          },
          {
            "name": "matrix",
            "in": "query",
            "description": "Select a matrix size that coresponds to the profile.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "96"
          },
          {
            "name": "mutationType",
            "in": "query",
            "description": "Enter mutation type",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "mutations",
            "in": "query",
            "description": "Enter mutations value",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_activity_options": {
      "get": {
        "tags": ["Mutational Activity"],
        "summary": "Retrieves Exposure options",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select which study to query data from.",
            "required": false,
            "schema": { "type": "string" },
            "value": "PCAWG"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) to filter data by.",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "signatureSetName",
            "in": "query",
            "description": "Enter signature set to filter data by.",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Enter cancer",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "signatureName",
            "in": "query",
            "description": "Enter signature",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_activity": {
      "get": {
        "tags": ["Mutational Activity"],
        "summary": "Retrieves Exposure data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select which study to query data from.",
            "required": true,
            "schema": { "type": "string" },
            "value": "PCAWG"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) to filter data by.",
            "required": true,
            "schema": { "type": "string" },
            "value": "WGS"
          },
          {
            "name": "signatureSetName",
            "in": "query",
            "description": "Enter signature set to filter data by.",
            "required": true,
            "schema": { "type": "string" },
            "value": "COSMIC_v3_Signatures_GRCh37_SBS96"
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Enter cancer",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "organ",
            "in": "query",
            "description": "Enter organ",
            "required": false,
            "schema": { "type": "string" },
            "value": ""
          },
          {
            "name": "sample",
            "in": "query",
            "description": "Enter sample",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "signatureName",
            "in": "query",
            "description": "Enter signature",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "exposure",
            "in": "query",
            "description": "Enter exposure value",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_signature_options": {
      "get": {
        "tags": ["Mutational Signature"],
        "summary": "Retrieves signature data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "description": "Enter source",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Reference_signatures", "Published_signatures"]
            },
            "value": "Reference_signatures"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "WGS"
          },
          {
            "name": "profile",
            "in": "query",
            "description": "Select profile",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["SBS", "DBS", "ID"]
            },
            "value": "SBS"
          },
          {
            "name": "matrix",
            "in": "query",
            "description": "Select matrix",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "signatureSetName",
            "in": "query",
            "description": "Enter signature set to filter data by.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "signatureName",
            "in": "query",
            "description": "Enter signature",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_signature": {
      "get": {
        "tags": ["Mutational Signature"],
        "summary": "Retrieves signature data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "description": "Enter source",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Reference_signatures", "Published_signatures"]
            },
            "value": "Reference_signatures"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["TS", "WES", "WGS"]
            },
            "value": "WGS"
          },
          {
            "name": "profile",
            "in": "query",
            "description": "Select profile",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["SBS", "DBS", "ID"]
            },
            "value": "SBS"
          },
          {
            "name": "matrix",
            "in": "query",
            "description": "Select matrix",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["96", "78", "83"]
            },
            "value": "96"
          },
          {
            "name": "signatureSetName",
            "in": "query",
            "description": "Enter signature set to filter data by.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "",
                "COSMIC_v3_Signatures_GRCh37_SBS96",
                "COSMIC_v3_Signatures_GRCh37_DBS78",
                "COSMIC_v3_Signatures_GRCh37_ID83"
              ]
            },
            "value": "COSMIC_v3_Signatures_GRCh37_SBS96"
          },
          {
            "name": "strandInfo",
            "in": "query",
            "description": "Enter strandInfo",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "strand",
            "in": "query",
            "description": "Enter strand",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "signatureName",
            "in": "query",
            "description": "Enter signature",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "mutationType",
            "in": "query",
            "description": "Enter mutationType",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "contribution",
            "in": "query",
            "description": "Enter contribution value",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_association_options": {
      "get": {
        "tags": ["Mutational Association"],
        "summary": "Retrieves association data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select profile",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "PCAWG"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "WGS"
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Enter cancer",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sample",
            "in": "query",
            "description": "Enter sample",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/mutational_association": {
      "get": {
        "tags": ["Mutational Association"],
        "summary": "Retrieves association data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "study",
            "in": "query",
            "description": "Select profile",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["PCAWG"]
            },
            "value": "PCAWG"
          },
          {
            "name": "strategy",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["TS", "WES", "WGS"]
            },
            "value": "WGS"
          },
          {
            "name": "cancer",
            "in": "query",
            "description": "Enter cancer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "value": "Biliary-AdenoCA"
          },
          {
            "name": "sample",
            "in": "query",
            "description": "Enter sample",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "icgcSpecimenId",
            "in": "query",
            "description": "Enter icgcSpecimenId",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "icgcDonorld",
            "in": "query",
            "description": "Enter icgcDonorld",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "dataSource",
            "in": "query",
            "description": "Enter dataSource",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "dataType",
            "in": "query",
            "description": "Enter dataType",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "variableName",
            "in": "query",
            "description": "Enter variableName",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "variableValue",
            "in": "query",
            "description": "Enter variableValue",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "variableValueType",
            "in": "query",
            "description": "Enter variableValueType",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },

    "/api/etiology": {
      "get": {
        "tags": ["Etiology"],
        "summary": "Retrieves etiology data",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Select profile",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "Cosmic"
          },
          {
            "name": "etiology",
            "in": "query",
            "description": "Select strategy (dataset) of the study.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": ""
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of records returned",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "10"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Records to skip",
            "required": false,
            "schema": {
              "type": "string"
            },
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    }
  }
}